Conflicts:
	parser/parser.y

+++ dom/tree.c
@@ -130,7 +130,7 @@ void red_black_tree_insert(tree_root* root, dom_node* node){
     y = x;
 
     if(compare(key(z), key(x)) == 0){
-      free(x->node);
+      destroy_dom_node(x->node);
       free(z);
       x->node = node;
       return;
@@ -195,6 +195,8 @@ static void __destroy_tree(tree_node* root){
 }
 
 void destroy_tree(tree_root* root){
+  if(root == NULL || root->root == NULL)
+    return;
   __destroy_tree(root->root);
   free(root);
 }
